# docs/swagger.yaml
basePath: /api
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.ErrorDetail'
    type: object
...
  api.ImportSummaryResponse:
    properties:
      data:
        $ref: '#/definitions/service.ImportSummary'
      meta: {}
    type: object
...
  service.DisciplineImportInput:
    properties:
      code:
        type: string
      description:
        type: string
      icon_url:
        type: string
      is_active:
        type: boolean
      metadata:
        type: object
      name:
        type: string
      team_size:
        type: integer
    type: object
  service.GamePlayerStatImportInput:
    properties:
      assists:
        type: integer
      damage_dealt:
        type: integer
      deaths:
        type: integer
      game_id:
        type: integer
      gold_earned:
        type: integer
      hero_name:
        type: string
      kills:
        type: integer
      player_id:
        type: integer
      team_id:
        type: integer
      was_mvp:
        type: boolean
    type: object
...
  service.MatchGameImportInput:
    properties:
      duration_seconds:
        type: integer
      game_number:
        type: integer
      had_technical_pause:
        type: boolean
      map_name:
        type: string
      match_id:
        type: integer
      pick_ban_phase:
        type: object
      score_team1:
        type: integer
      score_team2:
        type: integer
      started_at:
        type: string
      winner_team_id:
        type: integer
    type: object
  service.MatchImportInput:
    properties:
      format:
        type: string
      is_forfeit:
        type: boolean
      match_notes:
        type: object
      stage:
        type: string
      start_time:
        type: string
      team1_id:
        type: integer
      team2_id:
        type: integer
      tournament_id:
        type: integer
      winner_team_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: REST API for disciplines, teams, tournaments, matches, and reports.
  title: DB Course Project API
  version: "1.0"
paths:
  /batch-import/disciplines:
    post:
      consumes:
      - application/json
      parameters:
      - description: Disciplines to import
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/service.DisciplineImportInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import disciplines
      tags:
      - Utility
  /batch-import/disciplines/csv:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import disciplines from CSV
      tags:
      - Utility
  /batch-import/game-player-stats:
    post:
      consumes:
      - application/json
      parameters:
      - description: Game player stats to import
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/service.GamePlayerStatImportInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import game player stats
      tags:
      - Utility
  /batch-import/game-player-stats/csv:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import game player stats from CSV
      tags:
      - Utility
  /batch-import/match-games:
    post:
      consumes:
      - application/json
      parameters:
      - description: Match games to import
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/service.MatchGameImportInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import match games
      tags:
      - Utility
  /batch-import/match-games/csv:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import match games from CSV
      tags:
      - Utility
  /batch-import/matches:
    post:
      consumes:
      - application/json
      parameters:
      - description: Matches to import
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/service.MatchImportInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import matches
      tags:
      - Utility
  /batch-import/matches/csv:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import matches from CSV
      tags:
      - Utility
  /batch-import/players:
    post:
      consumes:
      - application/json
      parameters:
      - description: Players to import
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/service.PlayerImportInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import players
      tags:
      - Utility
  /batch-import/players/csv:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import players from CSV
      tags:
      - Utility
  /batch-import/squad-members:
    post:
      consumes:
      - application/json
      parameters:
      - description: Squad members to import
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/service.SquadMemberImportInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import squad members
      tags:
      - Utility
  /batch-import/squad-members/csv:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import squad members from CSV
      tags:
      - Utility
  /batch-import/team-profiles:
    post:
      consumes:
      - application/json
      parameters:
      - description: Team profiles to import
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/service.TeamProfileImportInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import team profiles
      tags:
      - Utility
  /batch-import/team-profiles/csv:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import team profiles from CSV
      tags:
      - Utility
  /batch-import/teams:
    post:
      consumes:
      - application/json
      parameters:
      - description: Teams to import
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/service.TeamImportInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Batch import teams
      tags:
      - Utility
...
swagger: "2.0"